+++ backend/srcs/src/channel/dto/index.ts
@@ -1,0 +1,4 @@
+export * from './createChannel.dto';
+export * from './createChannelPass.dto';
+export * from './muteChannelUser.dto';
+export * from './joinChannel.dto';