+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2023-10-03 - emadriga
+### Added
+- New module `channel` 
+
 ## 2023-10-01 - emadriga
 ### Fixed
 - Fix grouped date (month/day vs day/month) culture problems